+++ src/main/java/com/MoP2022/Team10/mapping/Map.java
@@ -101,7 +101,7 @@ public class Map extends OncePerRequestFilter {
     public ResponseEntity<ResDefault> addUserIngredient(
             @RequestParam(value = "userId")int userId,
             @RequestParam(value = "ingredientId")int ingredientId,
-            @RequestParam(value = "count")int count,
+            @RequestParam(value = "count")double count,
             @RequestParam(value = "expire")String expire
     ) {
         ResDefault res = new ResDefault();
